i2c funcionando

+++ eeprom.h
@@ -37,8 +37,10 @@ void I2C_SendACK(void);
 void I2C_SendNACK(void);
 uint8_t I2C_WriteByte(uint8_t data);
 uint8_t I2C_ReadByte(ACK_TypeDef SetValue);
-int I2C_Write(uint8_t addr, uint8_t* data, uint32_t len);
-int I2C_Read(uint8_t addr, uint8_t* data, uint32_t len);
+//int I2C_Write(uint8_t addr, uint8_t* data, uint32_t len);
+int I2C_Write(uint16_t addr, uint8_t* data, uint32_t len);
+//int I2C_Read(uint8_t addr, uint8_t* data, uint32_t len);
+int I2C_Read(uint16_t addr, uint8_t* data, uint32_t len);
 int I2C_WriteRepeated(uint8_t addr, uint8_t* data, uint32_t len);
 int I2C_ReadRepeated(uint8_t addr, uint8_t* data, uint32_t len);
